myapi/requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
build==1.2.2.post1
CacheControl==0.14.1
certifi==2024.12.14
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.7
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.9
dulwich==0.21.7
fastapi==0.115.6
fastjsonschema==2.21.1
filelock==3.16.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
installer==0.7.0
jaraco.classes==3.4.0
keyring==24.3.1
Mako==1.3.8
MarkupSafe==3.0.2
more-itertools==10.5.0
msgpack==1.1.0
packaging==24.2
pexpect==4.9.0
pkginfo==1.12.0
platformdirs==4.3.6
poetry==1.8.5
poetry-core==1.9.1
poetry-plugin-export==1.8.0
ptyprocess==0.7.0
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.10.1
PyMySQL==1.1.1
pyproject_hooks==1.2.0
pywin32-ctypes==0.2.3
RapidFuzz==3.11.0
redis==5.2.1
requests==2.32.3
requests-toolbelt==1.0.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tomlkit==0.13.2
trove-classifiers==2024.10.21.16
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.34.0
virtualenv==20.28.0
